DEV-37860; refactor: move fetch and sign-in to appropriate files
